+++ test/MultiSigWalletWithDailyLimit.js
@@ -56,7 +56,7 @@ contract('MultiSigWalletWithDailyLimit', (accounts) => {
             name: 'changeRequirement',
             type: 'function',
             inputs: [{
-                type: 'uint8',
+                type: 'uint256',
                 name: 'required'
             }]
         },
